2025-01-22 :: remove upbit setting

+++ deploy.sh
@@ -1,27 +1,28 @@
 #!/bin/bash
 
-IMAGE_NAME="autotrading:v1"
-BITHUMB_CONTAINER_NAME="bithumb"
+IMAGE_NAME="bithumb:v1"
+CONTAINER_NAME="bithumb"
 
-EXIST_BLUE=$(sudo docker-compose -p $BITHUMB_CONTAINER_NAME-blue -f docker-compose.blue.yaml ps | grep Up)
+EXIST_BLUE=$(sudo docker-compose -p $CONTAINER_NAME-blue -f docker-compose.blue.yaml ps | grep Up)
 IMAGE_ID=$(sudo docker images -q $IMAGE_NAME)
 
 if [ -z "$EXIST_BLUE" ]; then
   echo "blue up"
-  sudo docker stop upbit-green
-  sudo docker stop bithumb-green
-  sudo docker rm upbit-green
-  sudo docker rm bithumb-green
-  sudo docker-compose -p $BITHUMB_CONTAINER_NAME-blue -f docker-compose.blue.yaml up -d --build
+  sudo docker-compose -p $CONTAINER_NAME-blue -f docker-compose.blue.yaml up -d --build
+  BEFORE_COMPOSE_COLOR="green"
+  AFTER_COMPOSE_COLOR="blue"
 else
   echo "green up"
-  sudo docker stop upbit-blue
-  sudo docker stop bithumb-blue
-  sudo docker rm upbit-blue
-  sudo docker rm bithumb-blue
-  sudo docker-compose -p $BITHUMB_CONTAINER_NAME-green -f docker-compose.green.yaml up -d --build
+  sudo docker-compose -p $CONTAINER_NAME-green -f docker-compose.green.yaml up -d --build
+  BEFORE_COMPOSE_COLOR="blue"
+  AFTER_COMPOSE_COLOR="green"
 fi
 
 sleep 10
 
-sudo docker rmi "$IMAGE_ID"
+EXIST_AFTER=$(sudo docker-compose -p autotrading-${AFTER_COMPOSE_COLOR} -f docker-compose.${AFTER_COMPOSE_COLOR}.yaml ps | grep Up)
+if [ -n "$EXIST_AFTER" ]; then
+  sudo docker-compose -p autotrading-${BEFORE_COMPOSE_COLOR} -f docker-compose.${BEFORE_COMPOSE_COLOR}.yaml down
+  echo "$BEFORE_COMPOSE_COLOR down"
+  sudo docker rmi "$IMAGE_ID"
+fi
